refactor: interceptor, filter config 분리
