+++ src/main/java/kr/hhplus/be/server/config/FilterConfig.java
@@ -1,0 +1,24 @@
+package kr.hhplus.be.server.config;
+
+import jakarta.servlet.Filter;
+import kr.hhplus.be.server.filter.ApiLoggingFilter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@RequiredArgsConstructor
+@Configuration
+public class FilterConfig implements WebMvcConfigurer {
+
+	@Bean
+	public FilterRegistrationBean<Filter> logFilter() {
+		FilterRegistrationBean<Filter> filterRegistrationBean = new FilterRegistrationBean<>();
+		filterRegistrationBean.setFilter(new ApiLoggingFilter());
+		filterRegistrationBean.setOrder(1);
+		filterRegistrationBean.addUrlPatterns("/*");
+
+		return filterRegistrationBean;
+	}
+}
